Remove redundant includes and unused function
                                                                                                   is_neighbor_tile. Also fixed many signed/unsigned incompatibilities as well as integer size issues.

+++ src/grid_util.h
@@ -156,7 +156,8 @@ int is_corner_tile(struct surface_grid *sm, int idx);
 
 void find_shared_vertices_corner_tile_parent_wall(struct volume *world,
                                                   struct surface_grid *sm,
-                                                  int idx, int *shared_vert);
+                                                  int idx,
+                                                  long long int *shared_vert);
 
 void find_shared_vertices_for_neighbor_walls(struct wall *orig_wall,
                                              struct wall *nb_wall,
